fixes issue with picker jumping months
When the user has a day selected that is past
the number of days in the next month, the
picker would jump 2 months when the user
advances rather than 1.

This is because the month was set, then the
day was set, if the day number is `31` and
there are `28` days in a month, then js's
date class sets the month value forward.

Moving the day value being set before the
month prevents this behaivor. If the month
were to have been put forward, the setMonth
then resets it to the proper month.